Move BezGlyph for re-use
