+++ fontspector-checkapi/Cargo.toml
@@ -32,3 +32,5 @@ serde = { workspace = true }
 serde_json = "1.0"
 
 itertools = { workspace = true }
+
+kurbo = { version = "0.11", optional = true }
